requirements.txt
aiogram
python-dotenv
sympy